django/config/requirements.txt
amqp==5.1.0
asgiref==3.5.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cx-Oracle==8.3.0
Django==4.0.3
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.1.1
kombu==5.2.4
MarkupSafe==2.1.1
openapi-codec==1.3.2
pika==1.2.0
prompt-toolkit==3.0.28
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2022.1
requests==2.27.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
